+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Bug fixes
 
 * Display actual exception (if any) instead of SystemExitError for pow
+* Remote source always returns absolute path even if Torba.home_path/cache_path
+  is relative to current directory
 
 ## Version 0.1.1